lwNBDsvr: Small refactoring
