+++ modules/network/lwnbdsvr/drivers/atad.c
@@ -1,6 +1,6 @@
 #include "atad.h"
 
-int hdd_atad_init(struct nbd_context *me)
+int atad_init(struct nbd_context *me)
 {
     ata_devinfo_t *dev_info = ata_get_devinfo(0);
     if (dev_info != NULL && dev_info->exists) {
@@ -10,17 +10,17 @@ int hdd_atad_init(struct nbd_context *me)
     return 1;
 }
 
-int hdd_atad_read(struct nbd_context *me, void *buffer, uint64_t offset, uint32_t length)
+int atad_read(struct nbd_context *me, void *buffer, uint64_t offset, uint32_t length)
 {
     return ata_device_sector_io(0, buffer, (uint32_t)offset, length, ATA_DIR_READ);
 }
 
-int hdd_atad_write(struct nbd_context *me, void *buffer, uint64_t offset, uint32_t length)
+int atad_write(struct nbd_context *me, void *buffer, uint64_t offset, uint32_t length)
 {
     return ata_device_sector_io(0, buffer, (uint32_t)offset, length, ATA_DIR_WRITE);
 }
 
-int hdd_atad_flush(struct nbd_context *me)
+int atad_flush(struct nbd_context *me)
 {
     return ata_device_flush_cache(0);
 }
