+++ modules/network/lwnbdsvr/drivers/atad.h
@@ -1,5 +1,3 @@
-
-
 #ifndef ATAD_DRIVERS_NBD_H
 #define ATAD_DRIVERS_NBD_H
 
@@ -10,10 +8,10 @@
 extern "C" {
 #endif
 
-int hdd_atad_init(struct nbd_context *me);
-int hdd_atad_read(struct nbd_context *me, void *buffer, uint64_t offset, uint32_t length);
-int hdd_atad_write(struct nbd_context *me, void *buffer, uint64_t offset, uint32_t length);
-int hdd_atad_flush(struct nbd_context *me);
+int atad_init(struct nbd_context *me);
+int atad_read(struct nbd_context *me, void *buffer, uint64_t offset, uint32_t length);
+int atad_write(struct nbd_context *me, void *buffer, uint64_t offset, uint32_t length);
+int atad_flush(struct nbd_context *me);
 
 #ifdef __cplusplus
 }
