+++ modules/network/lwnbdsvr/lwNBD/nbd_server.h
@@ -1,15 +1,15 @@
 /****************************************************************/ /**
-                                                                    *
-                                                                    * @file nbd_server.h
-                                                                    *
-                                                                    * @author   Ronan Bignaux <ronan@aimao.org>
-                                                                    *
-                                                                    * @brief    Network Block Device Protocol implementation options
-                                                                    *
-                                                                    * Copyright (c) Ronan Bignaux. 2021
-                                                                    * All rights reserved.
-                                                                    *
-                                                                    ********************************************************************/
+ *
+ * @file nbd_server.h
+ *
+ * @author   Ronan Bignaux <ronan@aimao.org>
+ *
+ * @brief    Network Block Device Protocol implementation options
+ *
+ * Copyright (c) Ronan Bignaux. 2021
+ * All rights reserved.
+ *
+ ********************************************************************/
 
 /*
  * Redistribution and use in source and binary forms, with or without
@@ -42,6 +42,7 @@
 #define LWIP_HDR_APPS_NBD_SERVER_H
 
 #include "nbd-protocol.h"
+#include "nbd_protocol.h"
 #include "nbd_opts.h"
 
 //#include "lwip/apps/nbd_opts.h"
@@ -57,14 +58,22 @@
 //#include <errno.h>
 //#include <malloc.h>
 
-// TODO: Missing <byteswap.h> in PS2SDK
-//  pickup from https://gist.github.com/jtbr/7a43e6281e6cca353b33ee501421860c
+#ifdef DEBUG
+#define dbgprintf(args...) printf(args)
+#else
+#define dbgprintf(args...) \
+    do {                   \
+    } while (0)
+#endif
+
+//TODO: Missing <byteswap.h> in PS2SDK
+// pickup from https://gist.github.com/jtbr/7a43e6281e6cca353b33ee501421860c
 static inline uint64_t bswap64(uint64_t x)
 {
     return (((x & 0xff00000000000000ull) >> 56) | ((x & 0x00ff000000000000ull) >> 40) | ((x & 0x0000ff0000000000ull) >> 24) | ((x & 0x000000ff00000000ull) >> 8) | ((x & 0x00000000ff000000ull) << 8) | ((x & 0x0000000000ff0000ull) << 24) | ((x & 0x000000000000ff00ull) << 40) | ((x & 0x00000000000000ffull) << 56));
 }
 
-// TODO: Missing in PS2SK's "common/include/tcpip.h"
+//TODO: Missing in PS2SK's "common/include/tcpip.h"
 #if __BIG_ENDIAN__
 #define htonll(x) (x)
 #define ntohll(x) (x)
@@ -73,8 +82,8 @@ static inline uint64_t bswap64(uint64_t x)
 #define ntohll(x) bswap64(x)
 #endif
 
-// TODO: Missing in PS2SK's <stdint.h> , needed for "nbd-protocol.h"
-//  https://en.cppreference.com/w/c/types/integer
+//TODO: Missing in PS2SK's <stdint.h> , needed for "nbd-protocol.h"
+// https://en.cppreference.com/w/c/types/integer
 #define UINT64_MAX  0xffffffffffffffff
 #define UINT64_C(x) ((x) + (UINT64_MAX - UINT64_MAX))
 #endif
@@ -83,8 +92,7 @@ static inline uint64_t bswap64(uint64_t x)
 extern "C" {
 #endif
 
-// extern uint8_t buffer[];
-extern uint8_t nbd_buffer[NBD_BUFFER_LEN] __attribute__((aligned(64)));
+extern uint8_t nbd_buffer[];
 
 /** @ingroup nbd
  * NBD context containing callback functions for NBD transfers
@@ -94,49 +102,54 @@ extern uint8_t nbd_buffer[NBD_BUFFER_LEN] __attribute__((aligned(64)));
 struct nbd_context
 {
 
+    // move in
     char export_name[32];
-    char export_desc[64];
     uint64_t export_size; /* size of export in byte */
     uint16_t eflags;      /* per-export flags */
-    uint8_t blockshift;   /* in power of 2 for bit shifting - log2(blocksize) */
+
+    char export_desc[64];
+    uint8_t blockshift; /* in power of 2 for bit shifting - log2(blocksize) */
     uint8_t *buffer;
+
     /**
-     *  block device
-     * @param
-     * @returns
-     */
+   *  block device
+   * @param
+   * @returns
+   */
     int (*export_init)(struct nbd_context *me);
     /**
-     * Close block device handle
-     * @param handle File handle returned by open()
-     */
+   * Close block device handle
+   * @param handle File handle returned by open()
+   */
     //  void (*close)(struct nbd_context *me);
     /**
-     * Read from block device
-     * @param
-     * @param buffer Target buffer to copy read data to
-     * @param offset Offset in block to copy read data to
-     * @param length Number of blocks to copy to buffer
-     * @returns &gt;= 0: Success; &lt; 0: Error
-     */
+   * Read from block device
+   * @param
+   * @param buffer Target buffer to copy read data to
+   * @param offset Offset in block to copy read data to
+   * @param length Number of blocks to copy to buffer
+   * @returns &gt;= 0: Success; &lt; 0: Error
+   */
     int (*read)(struct nbd_context *me, void *buffer, uint64_t offset, uint32_t length);
     /**
-     * Write to block device
-     * @param me ()
-     * @param buffer Target buffer to copy write data to
-     * @param offset Offset in block to copy write data to
-     * @param length Number of blocks to copy to buffer
-     * @returns &gt;= 0: Success; &lt; 0: Error
-     */
+   * Write to block device
+   * @param me ()
+   * @param buffer Target buffer to copy write data to
+   * @param offset Offset in block to copy write data to
+   * @param length Number of blocks to copy to buffer
+   * @returns &gt;= 0: Success; &lt; 0: Error
+   */
     int (*write)(struct nbd_context *me, void *buffer, uint64_t offset, uint32_t length);
     /**
-     * Flush to block device
-     * @param me ()
-     * @returns &gt;= 0: Success; &lt; 0: Error
-     */
+   * Flush to block device
+   * @param me ()
+   * @returns &gt;= 0: Success; &lt; 0: Error
+   */
     int (*flush)(struct nbd_context *me);
 };
 
+#define nbd_send(a, b, c, d) lwip_send(a, b, c, d)
+int nbd_recv(int s, void *mem, size_t len, int flags);
 int nbd_init(struct nbd_context **ctx);
 
 #ifdef __cplusplus
